requirements.txt
# Stone-Sea - Document Generation Requirements
# For .docx template processing and generation

python-docx>=0.8.11
